refactor(auth): separate auth state and session logic
